This version implements rough mouse control, a different sized triangle, and a visual signal when the contains method is being called. The contain method is disabled for now, as it was causing more trouble than it was worth. We'll need to figure out something else.

+++ src/physics/LineSegment.java
@@ -77,7 +77,7 @@ public class LineSegment {
         Point origin = this.intersection(other);
         return new Ray(origin,R,speed);
     }
-         public int getYGivenX(int x){
+    public double getYGivenX(double x){
         double slope;
         double newY;
         if(this.a.x<this.b.x){
@@ -88,6 +88,6 @@ public class LineSegment {
             newY = this.b.y + slope*(x-this.b.x);
         }
         
-        return (int)newY;
+        return newY;
     }
 }
